// gen/models/new-notes.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface NewNotes
 */
export interface NewNotes {
    /**
     * 
     * @type {string}
     * @memberof NewNotes
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof NewNotes
     */
    body?: string;
    /**
     * 
     * @type {string}
     * @memberof NewNotes
     */
    author?: string;
    /**
     * 
     * @type {string}
     * @memberof NewNotes
     */
    published_at?: string;
    /**
     * 
     * @type {string}
     * @memberof NewNotes
     */
    created_by?: string;
    /**
     * 
     * @type {string}
     * @memberof NewNotes
     */
    updated_by?: string;
}