// gen/models/notes.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DOCUMENTATION
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: contact-email@something.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Notes
 */
export interface Notes {
    /**
     * 
     * @type {string}
     * @memberof Notes
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Notes
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof Notes
     */
    body?: string;
    /**
     * 
     * @type {object}
     * @memberof Notes
     */
    author?: object;
    /**
     * 
     * @type {string}
     * @memberof Notes
     */
    published_at?: string;
}